VBN Pipeline Testing Framework + New Pose Estimation + Improved Feature Detection

+++ include/apps/vbn/FeatureDetector.hpp
@@ -10,7 +10,8 @@ namespace vbn {
 // Configuration for the FeatureDetector (tunable parameters, no state).
 // ---------------------------------------------------------------------------
 struct FeatureDetectorConfig {
-    uint8_t BIN_THRESH        = 20;    // GRAY8 threshold for LED blobs (0..255)
+    uint16_t BIN_THRESH        = 20;      // Threshold in native DN units of the input image.
+    // No automatic scaling applied based on bit depth
     int     MIN_BLOB_AREA     = 100;      // reject tiny noise blobs [pixels]
     int     MAX_BLOB_AREA     = 20000;    // reject huge/glare blobs [pixels]
